refactor(database): emit UserPassword SETs without coalesce
Drop the patch/coalesce helpers so each typed update writes one SET.
Use uint8 for increment Delta so negatives are a compile error, and
let duplicate column assignments fail at the database.

+++ internal/domain/user_password_update.go
@@ -39,7 +39,7 @@ type UserPasswordLastSuccessfulCheckUpdate struct {
 func (*UserPasswordLastSuccessfulCheckUpdate) userPasswordUpdate() {}
 
 type UserPasswordIncrementFailedAttemptsUpdate struct {
-	Delta int16
+	Delta uint8
 }
 
 func (*UserPasswordIncrementFailedAttemptsUpdate) userPasswordUpdate() {}
